Preload PDF blob before download
